packepfecam/backend/requirements.txt
anyio==3.6.2
autopep8==2.0.1
bcrypt==4.0.1
beanie==1.17.0
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==3.4.8
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-model==0.0.5
MarkupSafe==2.1.2
motor==3.1.1
orjson==3.8.6
passlib==1.7.4
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.5
PyJWT==1.7.1
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.5
PyYAML==6.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
opencv-python==4.9.0.80
ultralytics==8.1.10
PyOpenGL==3.1.7
aiofiles==23.2.1
mongoengine==0.27.0
pytest==8.0.0
